+++ tests/testthat/test-messages.R
@@ -76,3 +76,32 @@ test_that("messages Test 4: Renamed variables messages are shown", {
     expect_message("Var . : '.*' was renamed to '.*'") %>%
     expect_message("Duplicate renamed term\\(s\\) were created")
 })
+
+# no_domain_log ----
+## Test 5: no_domain_log: No domain messages are shown ----
+test_that("messages Test 5: No domain messages are shown", {
+  # Remove empty lines in cli theme
+  local_cli_theme()
+
+  log_no_domain("adsl", "domains", "message") %>%
+    expect_message("Domain not found in metadata.") %>%
+    expect_message("Domain 'adsl' not found in metadata 'domains' column.")
+
+  adsl <- data.frame(
+    USUBJID = c(1001, 1002, 1003),
+    BRTHDT = c(1, 1, 2)
+  )
+
+  metadata <- data.frame(
+    dataset = c("adsl", "adsl"),
+    variable = c("USUBJID", "BRTHDT"),
+    order = c(1, 2)
+  )
+
+  xportr_order(adsl, metadata, "wrong_adsl", verbose = "message") %>%
+    expect_message("Domain not found in metadata.") %>%
+    expect_message("Domain 'wrong_adsl' not found in metadata 'dataset' column.") %>%
+    expect_message("2 variables not in spec and moved to end") %>%
+    expect_message("Variable moved to end in `.df`: `USUBJID` and `BRTHDT`") %>%
+    expect_message("All variables in dataset are ordered")
+})