+++ NEWS.md
@@ -1,5 +1,8 @@
 # xportr (development version)
 
+* Updated order messaging to clarify some messaging when all data in dataset is
+found in the specification. (#269)
+
 * `"hms"` was added to the default value of the `xportr.numeric_types` option.
 This ensures that `{xportr}` works smoothly with variables created by
 `admiral::derive_vars_dtm_to_tm()`. (#271)